requirements.txt
google-generativeai
youtube-transcript-api
PyGithub
scrapetube
